requirements.txt
absl-py==0.10.0
astor==0.8.1
astunparse==1.6.3
bleach==1.5.0
cached-property==1.5.2
cachetools==4.1.1
calmsize==0.1.3
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.33.1
h5py==2.10.0
html5lib==0.9999999
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgviz==1.2.2
importlib-metadata==2.0.0
imutils==0.5.3
joblib==0.17.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
labelme==4.5.6
lightgbm==3.0.0
Markdown==3.3.3
matplotlib==3.3.2
moviepy==1.0.3
mxnet==1.7.0.post1
networkx==2.5
nibabel==3.2.0
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.3
Pillow==8.0.1
proglog==0.1.9
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytorch-memlab==0.2.2
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.2.1
seaborn==0.11.0
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
terminaltables==3.1.0
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.4.0
torchstat==0.0.7
torchsummary==1.4
torchvision==0.5.0
tqdm==4.51.0
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
zipp==3.4.0
